refactor(product): AuctionApiClient 경로 변경으로 인한 수
